Merge GraphQL sources together in final operation
The final operation string now contains the operation _and_ fragment
definitions that it requires to run, if any. It is written in a way that
a single file can contain an operation and many fragment definitions.
You can also have a file with just multiple fragment definitions and no
operation, or a file with only an operation.